+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to the "mjml" extension will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+### [0.2.0] (2021-03-16)
+
+CSS property/value completions within `mj-style` tags.
+Attribute completion over multiple line.
+
 ### [0.1.4] (2021-03-14)
 
 Enhanced attribute completion description.